Fixed: incorrect calculation of DFS bytes read in metrics calculation in TaskHost.
git-svn-id: file:///mnt/c/Users/svgroot/Downloads/svn/repos/Jumbo/trunk@240 2f0f0f96-4b5b-11dd-99f4-89bc9257463c

+++ TaskHost/Program.cs
@@ -266,7 +266,7 @@ namespace TaskHost
                 metrics.RecordsWritten += output.RecordsWritten;
 
             if( taskConfig.DfsInput != null )
-                metrics.DfsBytesRead = inputStream.Position - (taskConfig.DfsInput.Block * _blockSize);
+                metrics.DfsBytesRead = inputStream.Position - (taskConfig.DfsInput.Block * (long)_blockSize);
             else
             {
                 FileInputChannel fileInputChannel = inputChannel as FileInputChannel;
